Move MCP plugin helpers into plugins service
Interface diffs:

- import "github.com/valon-technologies/gestalt/server/internal/composite"

+ import "github.com/valon-technologies/gestalt/server/services/plugins/composite"

- import "github.com/valon-technologies/gestalt/server/internal/mcp"

+ import "github.com/valon-technologies/gestalt/server/services/plugins/mcp"

- import "github.com/valon-technologies/gestalt/server/internal/mcpoauth"

+ import "github.com/valon-technologies/gestalt/server/services/plugins/mcpoauth"

Examples:

provider := composite.New(name, apiProvider, mcpUpstream)

server := mcp.NewServer(mcp.Config{Providers: providers, Broker: broker})

handler := mcpoauth.NewHandler(mcpoauth.HandlerConfig{MCPURL: url, RedirectURL: redirectURL})
