+++ gestaltd/services/plugins/composite/merged.go
