+++ gestaltd/services/plugins/composite/merged_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/core/catalog"
-	"github.com/valon-technologies/gestalt/server/internal/composite"
+	"github.com/valon-technologies/gestalt/server/services/plugins/composite"
 )
 
 type fakeProvider struct {
