+++ gestaltd/services/plugins/composite/operation_resolution.go
