+++ gestaltd/services/plugins/composite/provider.go
