+++ gestaltd/services/plugins/mcp/binding.go
