+++ gestaltd/services/plugins/mcp/binding_test.go
@@ -13,15 +13,15 @@ import (
 	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/core/catalog"
 	coretesting "github.com/valon-technologies/gestalt/server/core/testing"
-	"github.com/valon-technologies/gestalt/server/internal/composite"
 	"github.com/valon-technologies/gestalt/server/internal/config"
 	"github.com/valon-technologies/gestalt/server/internal/coredata"
-	gestaltmcp "github.com/valon-technologies/gestalt/server/internal/mcp"
 	"github.com/valon-technologies/gestalt/server/internal/testutil"
 	"github.com/valon-technologies/gestalt/server/services/authorization"
 	"github.com/valon-technologies/gestalt/server/services/identity/principal"
 	"github.com/valon-technologies/gestalt/server/services/invocation"
+	"github.com/valon-technologies/gestalt/server/services/plugins/composite"
 	coreintegration "github.com/valon-technologies/gestalt/server/services/plugins/declarative"
+	gestaltmcp "github.com/valon-technologies/gestalt/server/services/plugins/mcp"
 	"github.com/valon-technologies/gestalt/server/services/plugins/mcpupstream"
 
 	mcpgo "github.com/mark3labs/mcp-go/mcp"
