+++ gestaltd/services/plugins/mcp/handlers.go
