+++ gestaltd/services/plugins/mcp/projection.go
