+++ gestaltd/services/plugins/mcp/session_tools.go
